% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{assign_expression_data}
\alias{assign_expression_data}
\title{assign expression data to the ortholog matrix}
\usage{
assign_expression_data(x, meta_table, file_dir = ".")
}
\arguments{
\item{x}{Data with information for paired 1-to-1 orthologous pairs. It should be the
list generated using \code{\link{ortholog_match}}.}

\item{meta_table}{The metatable with the details for the expression data of all involved
species. This table should be of csv format.}

\item{file_dir}{The directory that containing the expression data files. Default: "."}
}
\value{
Matrix, with rows as GeneID or GeneName, and columns as species. This matrix
can be used for subsequent analysis and visualization.
}
\description{
Compare between the ortholog matrix and expression data, hence obtain the
expression matrix for the ortholog genes across all compared species.
Theoretically, this function can also be used to integrate other types of
data apart from the expression files.
}
\examples{
# get ortholog matching information
species5.orth <- ortholog_match(
  species_list = c(
  "human", "mouse", "cattle", "pig", "opossum"
  )
)

# extract ortholog expression matrix for the compared species
species5.expr <- assign_expression_data(
  x = species5.orth,
  meta_table = "expression_metatable.csv",
  file_dir   = "inst/extdata"
)

# check top lines of the expression matrix
head(Species5.expr)

# write the expression matrix to a file
write.table(
  species5.expr,
  file = "Species5_expr.txt",
  sep = "\t", quote = FALSE,
   row.names = TRUE, col.names = TRUE
)
 
}
