% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{list_species}
\alias{list_species}
\title{List all supported species}
\usage{
list_species(pattern)
}
\arguments{
\item{pattern}{The pattern to filtering the species list. The pattern will be searched
against the common and scientific names, with the information for matched
species returned. To be noted, the matching is case insensitive.}
}
\value{
Data frame, with the information for supported species
}
\description{
Only species with genome available in ENSEMBL database are supported.
This function will show the information for all these species, or those filtered
by the user specified pattern.
}
\examples{
# list all species based on locally saved species list file
list_species()

# list all species with common or scientific name matching given pattern
list_species(pattern = "human")
list_species(pattern = "sapiens")
list_species(pattern = "macaque")

}
