% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{list_supported_species}
\alias{list_supported_species}
\title{List all supported species from ENSEMBL database}
\usage{
list_supported_species(pattern)
}
\arguments{
\item{pattern}{The pattern to filtering the species list. The pattern will be searched
against the common and scientific names, with the information for matched
species returned. To be noted, the matching is case insensitive.}
}
\value{
Data frame, with the information for supported species
}
\description{
Only species with genome available in ENSEMBL database are supported.
This function will show the information for all these species, or those
filtered by the user specified pattern.
}
\examples{
## list all species based on locally saved species list file
list_supported_species()

## list all species with common or scientific name matching given pattern
# human
list_supported_species(pattern = "human")
list_supported_species(pattern = "sapiens")

# macaque
list_supported_species(pattern = "macaque")

}
