% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ortholog.R
\name{ortholog_match}
\alias{ortholog_match}
\title{Match the 1-to-1 orthologous genes across multiple species}
\usage{
ortholog_match(species_list, species_anchor, host)
}
\arguments{
\item{species_list}{A vector for all the species to be analyzed. Should be common names such as
"human", "mouse", etc. Use \code{\link{list_supported_species}} to get the
names of all supported species. This parameter is case-insensitive.}

\item{species_anchor}{A species name from the provided species_list. This species will server as
the anchor to be compared against by all other species. Default:The first
species from the provided species_list.}

\item{host}{ENSEMBL host (mirror) to connect to. Default: https://www.ensembl.org. By
default, bioMart uses the latest releases of genome assemblies from the
ENSEMBL database, which may not match the genome version of given data
sometimes. Earlier releases can be used by specifying an archieved host
address, which can be  found from the ENSEMBL website.}
}
\value{
A matrix with the information for 1-to-1 ortholog pairs. The list has
corresponding species data frames (each for one species) with columns GeneID,
GeneName, Chrom,GeneType. Additonal attributes including Species and
SpeciesAbbr are also attached to the return list.
}
\description{
Ortholog matching is performed based on the ortholog information retrieved
from ENSEMBL. The R package biomaRt is used to access ENSEMBL.To work
properly, make sure biomaRt package is installed and the internet connection
works. It may take sometime to download data from ENSEMBL - depending on the
network speed. Only species with their genomes availale in ENSEMBL database
are supported. Use \code{\link{list_supported_species}} to get the
information for all the supported species.
}
\examples{
## Get 1:1 orthologs between any numbers of supplied species

# Two species
hs2mm.orth <- ortholog_match(
  species_list = c("human", "mouse")
)

# Three species
hs2mm2ss.orth <- ortholog_match(
  species_list = c("human", "mouse", "pig")
)

# Ten species
species5.orth <- ortholog_match(
  species_list = c(
  "human", "mouse", "cattle", "pig", "opossum"
  )
)

## Specify mouse, instead of human, as the anchor species
hs2mm2ss.orth <- ortholog_match(
  species_list = c("human", "mouse", "pig"),
  species_anchor = "mouse"
)

## Use different host (mirrors) for ENSEMBL
hs2mm2ss.orth <- ortholog_match(
  species_list = c("human", "mouse", "pig"),
  host = "https://www.ensembl.org"
)

hs2mm2ss.orth <- ortholog_match(
  species_list = c("human", "mouse", "pig"),
  host = "https://useast.ensembl.org"
)

hs2mm2ss.orth <- ortholog_match(
  species_list = c("human", "mouse", "pig"),
  host = "https://asia.ensembl.org"
)

## Use archived host for ENSEMBL, which contains old genome releases
hs2mm2ss.orth <- ortholog_match(
  species_list = c("human", "mouse", "pig"),
  host = "https://nov2020.archive.ensembl.org"
)
}
