% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_correlation_heatmap}
\alias{plot_correlation_heatmap}
\title{plot the correlation heatmap based on expression matrix}
\usage{
plot_correlation_heatmap(
  x,
  method = "p",
  file_name = "cor_heatmap.svg",
  width = 6,
  height = 5,
  pointsize = 12,
  ...
)
}
\arguments{
\item{x}{Ortholog expression matrix obtained by using \code{\link{assign_expression_data}}.}

\item{method}{The method for calculating correlation efficiency. It can be "p" for Pearson,
or "s" for Spearman. Default: "p"}

\item{file_name}{The file name for the generated figure. Default: "cor_heatmap.svg"}

\item{width}{The width for the generated figure. Default: 6.}

\item{height}{The height for the generated figure. Default: 5.}

\item{pointsize}{The point size for the generated figure. Default: 12}
}
\description{
Plot the correlation heatmap based on the expression matrix of the ortholog
genes across multiple species.
}
\examples{
# plot the correlation heatmap of SVG format
plot_correlation_heatmap(
  x = species5.expr,
  method     = "s",
  file_name  = "cor_heatmap.svg",
  width      = 6,
  height     = 5,
  pointsize  = 12
)

}
