% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_PCA_biplot}
\alias{plot_PCA_biplot}
\title{plot the PCA biplot based on expression matrix}
\usage{
plot_PCA_biplot(
  x,
  file_name = "PCA_biplot.svg",
  width = 5,
  height = 3,
  pointsize = 12,
  ...
)
}
\arguments{
\item{x}{Ortholog expression matrix obtained by using \code{\link{assign_expression_data}}.}

\item{file_name}{The file name for the generated figure. Default: "PCA_biplot.svg"}

\item{width}{The width for the generated figure. Default: 5.}

\item{height}{The height for the generated figure. Default: 4.5.}

\item{pointsize}{The point size for the generated figure. Default: 12}
}
\description{
Plot the PCA biplot based on the expression matrix of the ortholog
genes across multiple species.
}
\examples{
# plot the PCA biplot of SVG format
plot_PCA_biplot(
  x = species5.expr,
  file_name  = "PCA_biplot.svg",
  width      = 5,
  height     = 4.5,
  point_size = 12
)

}
