% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{verify_ortholog_data}
\alias{verify_ortholog_data}
\title{Verify if the orthologue matching data is normal}
\usage{
verify_ortholog_data(x)
}
\arguments{
\item{x}{Data with information for paired 1-to-1 orthologous pairs. It should be the
list generated using \code{\link{ortholog_match}}.}
}
\value{
List, unchanged/updated data with the same structre as the input data x.
}
\description{
Examine given data to see if its format is normal. Further, if no "Species"
attribute is present, try to set as Species_1, Species_2 etc, and return
the updated data. The species name is required for some other functions.
}
\examples{
# get ortholog matching information
hs2mm.orth <- ortholog_match(c("human", "mouse"))

# verify the obtained ortholog matching information
hs2mm.orth <- verify_ortholog_data(hs2mm.orth)

}
