% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic.R
\name{summarize_ortholog_gene}
\alias{summarize_ortholog_gene}
\title{Summarize the orthologues for each genetype or chromosome}
\usage{
summarize_ortholog_gene(x, group = c("genetype", "chrom"))
}
\arguments{
\item{x}{Data frame with ortholog annotations for corresponding species. It is generated
using ortholog_match(species_list=c()).}

\item{group}{genetype or chrom, the number of genes for each group based on specified
tags will be listed.}
}
\value{
Data frame with the count of genes for each genetype or chrom in each species.
}
\description{
Summarize the orthologues for each genetype or chromosome
}
\examples{
hs2mm2ss.orth <- ortholog_match(species_list=c("human", "mouse","pig"))
summarize_ortholog_gene(hs2mm2ss.orth, "genetype")

}
